fix(drift-check): split docs-claim tests under FR-006 cap, pin fragment scan
Review feedback on #2239 (F01-F05, F07; F06 deferred for maintainer policy):

- F01: move docs numeric-claim + lane-cap describes into
  drift-check-docs-claims.test.ts (422 + 343 lines, both under the 500 cap)
- F02/F04: fragment scan uses matchAll over /g-cloned regexes so every
  occurrence is checked, and dedupes per (file, wrong number) so overlapping
  phrasings collapse to one finding while distinct numbers stay separate
- F03: fixtures interpolate QA_GATE_PIPELINE_STEP_COUNT instead of
  hand-copied 15/12 literals
- F05: cover pending-fragment filters (non-.md files and subdirectories
  under docs/releases/pending are skipped)
- new regression pins: stale second occurrence of the same phrasing is
  flagged; overlapping phrasings at one wrong number yield one finding
- F07: invariant row wording now matches actual spelled-form coverage
